feat: in-house trial (#23)
Closes #18

Moving trials from stripe to in-house and doing it ourselves since our
mental models don't align

## Old Text

The current issue is that when I end the trial, it fails since the user
doesn't have a default payment option - even though they've added one.
Therefore Stripe throws a 400 cause you can't update the subscription to
expire the trial without being able to invoice the customer (which they
haven't given consent to)

Not sure if our use-case is even possible tbf, the response posted in
the SO post doesn't work, and even if you could make it work its really
bad UX - since it doesn't say that we'll invoice you when in the setup
mode, so you'll add your card info without really knowing what's going
to happen.

The main challenge is that stripe doesn't allow you to cancel a trial
immediately from the checkout session, n'or do they let you update a
current subscription. The webhooks we can hook into is also not the
best, and doesn't provide the correct details. I'd think that this could
be for security reasons and that apps shouldn't abuse the system and do
too crazy stuff?

Stripe's mental model is that trials are fully valid subscriptions, when
in our head it's really not. I think we either need to roll our own
trial or shift our mental model to be more aligned with Stripe. This
means we'd embrace trials being valid subscriptions and stop treating
them differently, i.e. if I setup my card for future payments I wont be
invoiced until the end of my trial.

A different approach I think you could also make work is having the
Trial being it's separate item with a separate price_id, and giving that
for free for 7 days - but that wouldn't allow for a seamless upgrade
when the trial would end.

Or, since we don't treat trials and subscriptions the same, we can also
just cancel the trial immediately, and create a new subscription. Sure,
might result in an extra notification but as a user I wouldn't be too
confused if I got 1 email that my trial ends and a new subscription has
been initialized, especially since that's exactly what we'd be doing